requirements.txt
jupyter==1.1.1
streamlit==1.45.0
nbformat==5.10.4
nbconvert==7.16.6
pandas==2.2.3
matplotlib==3.10.3
altair==5.5.0
plotly-express==0.4.1
folium==0.20.0
